move extension planner
