+++ rust/sedona-spatial-join/tests/spatial_join_integration.rs
@@ -51,6 +51,7 @@ use sedona_spatial_join::{
     register_planner, spatial_predicate::RelationPredicate, ProbeShuffleExec, SpatialJoinExec,
     SpatialPredicate,
 };
+use sedona_spatial_join_common::extension_planner::SpatialJoinExtensionPlanner;
 use sedona_spatial_join_common::query_planner::SedonaQueryPlanner;
 use sedona_testing::datagen::RandomPartitionedDataBuilder;
 use tokio::sync::OnceCell;
@@ -155,11 +156,13 @@ fn setup_context(options: Option<SpatialJoinOptions>, batch_size: usize) -> Resu
     session_config = add_sedona_option_extension(session_config);
     let mut state_builder = SessionStateBuilder::new();
     if let Some(options) = options {
-        let (sb, ext_planner) = register_planner(state_builder)?;
+        let (sb, factory) = register_planner(state_builder)?;
         state_builder = sb;
 
         state_builder = state_builder.with_query_planner(Arc::new(
-            SedonaQueryPlanner::new(vec![ext_planner]),
+            SedonaQueryPlanner::new(vec![Arc::new(
+                SpatialJoinExtensionPlanner::new(vec![factory]),
+            )]),
         ));
         let opts = session_config
             .options_mut()
